Restructure project:   - Renamed valueChanger to startValue to make more sense of its scope   - Switched Count from int to an int64 value   - Separated components from main     - Created a directory named components, which contains counter and startValue components       - StartValue is the state handler, which takes:          - an anonymous struct prop to valueInput widget.Editor            - currently, the invalid prop of type bool has no use          - the widget.Clickable meant to change the ResetVal and displayed number   - Separated global variables from main to a new directory named globals:     - This handles the values of Count and ResetVal       - ResetVal is now used by the Reset Button widget to change to the changed starting value   - Some fixes:     - Readjusted the startValue component layout Flex weight from 0.1 to 0.112 !!!     - Added a dark-green color.NRGBA value to make the displayed number from the header more readable when displaying positive numbers
